share with elementwise horizontal op
